+++ src/getDirectory.js
@@ -1,8 +1,7 @@
 // @flow
-'use strict';
 
-const path = require('path');
-const isDirectory = require('is-directory');
+import path from 'path';
+import isDirectory from 'is-directory';
 
 function getDirectory(filepath: string): Promise<string> {
   return new Promise((resolve, reject) => {
@@ -19,4 +18,4 @@ getDirectory.sync = function getDirectorySync(filepath: string): string {
   return isDirectory.sync(filepath) ? filepath : path.dirname(filepath);
 };
 
-module.exports = getDirectory;
+export { getDirectory };